Вынес fill_credentials_and_refresh в базовые методы. Перенес логин/пароль в .env
