+++ pages/httpwatch_auth_page.py
@@ -1,4 +1,3 @@
-import base64
 
 import allure
 from dotenv import load_dotenv
@@ -27,17 +26,6 @@ class HttpWatchAuthPage(BasePage):
     @allure.step('Нажать кнопку "Display Image"')
     def click_display_image_button(self):
         self.click_button(self.DISPLAY_IMAGE_BUTTON)
-
-    @allure.step('Ввести значения логина и пароля, обновить страницу')
-    def fill_credentials_and_refresh(self, username, password):
-        credentials = f'{username}:{password}'
-        encoded_credentials = base64.b64encode(credentials.encode()).decode('ascii')
-
-        self.browser.execute_cdp_cmd('Network.enable', {})
-        self.browser.execute_cdp_cmd('Network.setExtraHTTPHeaders', {
-            'headers': {'Authorization': f'Basic {encoded_credentials}'}
-        })
-        self.browser.refresh()
 
     @allure.step('Проверить отображение изображения на странице')
     def check_image_is_visible(self):
